:: MAZE solving problem ::
Checked in the code for solving the problem of finding a way out from
the maze. The maze is basically a m x n grid having some open cells and
walls in other cells. A robot starts at one point and need to reach out
to a destination target. The solution returns true if a path exists in
the maze. Otherwise it returns false.